requirements.txt
Flask==3.1.0
psycopg2==2.9.10
pytz==2024.1
Requests==2.32.3
Werkzeug==3.1.3
Flask-Session==0.8.0
python-dotenv==1.0.1